// React/app/GerAdv_TS/NECompromissos/Interfaces/interface.NECompromissos.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface INECompromissos {
// 202501251
    id: number;
 
	tipocompromisso: number,
	palavrachave: number,
	provisionar: boolean,
	textocompromisso: string,
}

export interface INECompromissosFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface INECompromissosIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}